+++ docs/changelog.md
@@ -37,6 +37,9 @@
     * Cog Machines can now be disabled with a redstone signal [(#145)](https://github.com/codetaylor/pyrotech/issues/145)
     * New Item: Lump of Refractory Clay <pyrotech:material:35>
     * Terracotta and Glazed Terracotta recipes have been added to the Pit Kiln, Stone Kiln, and Refractory Kiln [(#134)](https://github.com/codetaylor/pyrotech/issues/134)
+    * Config options to change the input capacity of the Bloomery and Wither Forge [(#79)](https://github.com/codetaylor/pyrotech/issues/79)
+        * `module.tech.Bloomery.cfg` -> `BLOOMERY` -> `CAPACITY`
+        * `module.tech.Bloomery.cfg` -> `WITHER_FORGE` -> `CAPACITY`
 
   * **Fixed:**
     * Pit Kiln only plays wood placement sound for the last log placed